DEV: modularize parsers
